+++ gDRplots/NEWS.md
@@ -1,5 +1,6 @@
 ## gDRplots 0.0.18 - 2024-07-26
 * improve the ratio of square plots
+* add a function for finding proper render size for plot
 
 ## gDRplots 0.0.17 - 2024-07-25
 * add a function for saving plots